feat: add dedicated Dashboard and Behavior settings screens
- Introduced new Dashboard Overview and Behavior settings screens, enhancing user navigation and organization within the app.
- Updated MainScreen and SettingsScreen to include routes and links to the new settings, improving accessibility.
- Enhanced CHANGELOG.md to document the addition of these features and their functionalities.
